699: Moving FAPI filter to v1 subpackage
FAPI v2 is in the works, so it makes sense to place classes that are version specific in a version subpackage.

https://github.com/SecureApiGateway/SecureApiGateway/issues/699
